+++ packages/fe-fpm-writer/test/unit/page/custom.test.ts
@@ -8,6 +8,8 @@ import { generateCustomPage, validateBasePath } from '../../../src';
 import { FCL_ROUTER } from '../../../src/common/defaults';
 import { detectTabSpacing } from '../../../src/common/file';
 import { tabSizingTestCases } from '../../common';
+import type { Logger } from '@sap-ux/logger';
+import { i18nNamespaces, translate } from '../../../src/i18n';
 
 describe('CustomPage', () => {
     const testDir = '' + Date.now();
@@ -298,7 +300,6 @@ describe('CustomPage', () => {
                 ...input,
                 pageBuildingBlockTitle: 'Test Page Title'
             };
-
             fs.writeJSON(join(target, 'webapp/manifest.json'), testManifestWithNoRouting);
             await generateCustomPage(target, inputWithPageBuildingBlockTitle, fs);
 
@@ -309,6 +310,33 @@ describe('CustomPage', () => {
             expect(viewXml).toContain('Test Page Title');
 
             expect(fs.read(join(target, 'webapp/ext/customPage/CustomPage.view.xml'))).toMatchSnapshot();
+        });
+
+        test('should log a warning when min ui5 version is not met for page building block feature', async () => {
+            const target = join(testDir, 'single-page-no-fcl');
+            const t = translate(i18nNamespaces.buildingBlock, 'pageBuildingBlock.');
+            const inputWithPageBuildingBlockTitle = {
+                ...input,
+                minUI5Version: '1.120',
+                pageBuildingBlockTitle: 'Test Page Title'
+            };
+            fs.writeJSON(join(target, 'webapp/manifest.json'), testManifestWithNoRouting);
+
+            const log = { warn: jest.fn() } as unknown as Logger;
+
+            await generateCustomPage(target, inputWithPageBuildingBlockTitle, fs, log);
+
+            expect(log.warn).toHaveBeenCalledWith(
+                t('minUi5VersionRequirement', { minUI5Version: inputWithPageBuildingBlockTitle.minUI5Version })
+            );
+
+            // page macros should not be added
+            const viewXmlPath = join(target, 'webapp/ext/customPage/CustomPage.view.xml');
+            expect(fs.exists(viewXmlPath)).toBe(true);
+            const viewXml = fs.read(viewXmlPath).toString();
+            expect(viewXml).not.toContain('macros:Page');
+            expect(viewXml).not.toContain('Test Page Title');
+            expect(viewXml).toContain('<Page');
         });
     });