+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.7.6
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.6.4
+-   @sap-ux/odata-service-inquirer@2.7.2
+
 ## 0.7.5
 
 ### Patch Changes